[insn][jal][execenv] fix execution env instructions, add JAL and JALR

+++ include/registers.hpp
@@ -3,7 +3,7 @@
 
 namespace rv32i_sim {
 
-enum class Register : uint8_t {
+enum class Register : uint32_t {
   X0 = 0, X1 = 1, X2 = 2, X3 = 3,
   X4 = 4, X5 = 5, X6 = 6, X7 = 7,
   X8 = 8, X9 = 9, X10 = 10, X11 = 11,
